refactor service to allow unauthenticated proxy
