+++ terraform/packages-service.yml
@@ -65,16 +65,16 @@ paths:
         '5XX':
           $ref: '#/components/responses/Error'
 
-  /s3/proxy:
+  /presign/s3:
     get:
-      summary: Proxy for S3 package file requests
+      summary: Generate presigned URLs for S3 objects
       description: |
-        Generates presigned URLs and proxies requests to S3 with proper CORS handling.
-        This endpoint handles GET requests for package files and returns appropriate CORS headers
-        for clients that require OPTIONS preflight support (e.g., DuckDB, web browsers).
+        Authenticated endpoint that generates presigned URLs for S3 objects.
+        This endpoint validates dataset and package ownership before generating URLs.
+        Supports both package files and viewer assets.
       x-amazon-apigateway-integration:
         $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
-      operationId: s3ProxyGet
+      operationId: presignS3Get
       security:
         - token_dataset_auth: [ ]
       tags:
@@ -131,13 +131,13 @@ paths:
         '5XX':
           $ref: '#/components/responses/Error'
     options:
-      summary: CORS preflight for S3 package file requests
+      summary: CORS preflight for presigned URL generation
       description: |
         Handles OPTIONS preflight requests for CORS support.
-        Returns appropriate CORS headers to allow clients to access S3 package files.
+        Returns appropriate CORS headers for the presign endpoint.
       x-amazon-apigateway-integration:
         $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
-      operationId: s3ProxyOptions
+      operationId: presignS3Options
       security:
         - token_dataset_auth: [ ]
       tags:
@@ -176,13 +176,13 @@ paths:
         '5XX':
           $ref: '#/components/responses/Error'
     head:
-      summary: HEAD request proxy for S3 package file metadata
+      summary: HEAD request for S3 package file metadata
       description: |
-        Generates presigned URLs and proxies HEAD requests to S3 to retrieve metadata.
+        Authenticated endpoint that generates presigned URLs and retrieves S3 metadata.
         Clients like DuckDB use HEAD requests to get Content-Length before making range requests.
       x-amazon-apigateway-integration:
         $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
-      operationId: s3ProxyHead
+      operationId: presignS3Head
       security:
         - token_dataset_auth: [ ]
       tags:
@@ -224,6 +224,121 @@ paths:
                 type: string
         '4XX':
           $ref: '#/components/responses/Unauthorized'
+        '5XX':
+          $ref: '#/components/responses/Error'
+
+  /proxy/s3:
+    get:
+      summary: Proxy S3 requests using presigned URLs
+      description: |
+        Unauthenticated proxy endpoint for S3 requests using presigned URLs.
+        This endpoint accepts a presigned URL as input and proxies the request,
+        handling CORS requirements for clients like DuckDB that cannot add custom headers.
+        GET requests return HTTP redirects to the presigned URL.
+      x-amazon-apigateway-integration:
+        $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
+      operationId: proxyS3Get
+      tags:
+        - Packages Service
+      parameters:
+        - in: query
+          name: presigned_url
+          schema:
+            type: string
+          required: true
+          description: The presigned S3 URL to proxy
+      responses:
+        '307':
+          description: Temporary redirect to the presigned URL
+          headers:
+            Location:
+              schema:
+                type: string
+            Access-Control-Allow-Origin:
+              schema:
+                type: string
+        '4XX':
+          $ref: '#/components/responses/BadRequest'
+        '5XX':
+          $ref: '#/components/responses/Error'
+    options:
+      summary: CORS preflight for S3 proxy requests
+      description: |
+        Handles OPTIONS preflight requests for the unauthenticated proxy endpoint.
+        Returns appropriate CORS headers to allow clients to access S3 through the proxy.
+      x-amazon-apigateway-integration:
+        $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
+      operationId: proxyS3Options
+      tags:
+        - Packages Service
+      parameters:
+        - in: query
+          name: presigned_url
+          schema:
+            type: string
+          required: true
+          description: The presigned S3 URL (required for consistency but not used in OPTIONS)
+      responses:
+        '204':
+          description: Successful CORS preflight response
+          headers:
+            Access-Control-Allow-Origin:
+              schema:
+                type: string
+            Access-Control-Allow-Methods:
+              schema:
+                type: string
+            Access-Control-Allow-Headers:
+              schema:
+                type: string
+            Access-Control-Max-Age:
+              schema:
+                type: string
+        '4XX':
+          $ref: '#/components/responses/BadRequest'
+        '5XX':
+          $ref: '#/components/responses/Error'
+    head:
+      summary: HEAD request proxy for S3 metadata
+      description: |
+        Unauthenticated proxy endpoint that forwards HEAD requests to S3 using presigned URLs.
+        Returns S3 metadata headers directly (not a redirect) for clients like DuckDB.
+      x-amazon-apigateway-integration:
+        $ref: '#/components/x-amazon-apigateway-integrations/packages-service'
+      operationId: proxyS3Head
+      tags:
+        - Packages Service
+      parameters:
+        - in: query
+          name: presigned_url
+          schema:
+            type: string
+          required: true
+          description: The presigned S3 URL to retrieve metadata from
+      responses:
+        '200':
+          description: Successful HEAD response with S3 metadata
+          headers:
+            Content-Length:
+              schema:
+                type: string
+            Content-Type:
+              schema:
+                type: string
+            ETag:
+              schema:
+                type: string
+            Last-Modified:
+              schema:
+                type: string
+            Accept-Ranges:
+              schema:
+                type: string
+            Access-Control-Allow-Origin:
+              schema:
+                type: string
+        '4XX':
+          $ref: '#/components/responses/BadRequest'
         '5XX':
           $ref: '#/components/responses/Error'
